Consolidate SHA-256 implementations into shared crypto utility
